[CP-1090] Implementing infinity loading for large contacts lists
