+++ packages/app/package.json
@@ -165,6 +165,7 @@
     "@types/react-beautiful-dnd": "^13.1.2",
     "@types/react-dom": "^17.0.11",
     "@types/react-hot-loader": "^4.1.1",
+    "@types/react-infinite-scroller": "^1.2.3",
     "@types/react-modal": "^3.13.1",
     "@types/react-redux": "^7.1.24",
     "@types/react-router": "^5.1.18",
@@ -253,6 +254,7 @@
     "react-google-button": "^0.7.2",
     "react-hook-form": "~7.12.2",
     "react-hot-loader": "^4.13.0",
+    "react-infinite-scroller": "^1.2.6",
     "react-intersection-observer": "^9.3.5",
     "react-intl": "5.0.2",
     "react-intl-translations-manager": "^5.0.3",
