+++ packages/app/src/contacts/components/contact-group/index.ts
@@ -3,5 +3,4 @@
  * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
  */
 
-export * from "./contact-category.object"
-export * from "./contact.object"
+export * from "./contact-group.component"
