+++ packages/app/src/contacts/selectors/contact-list.selector.ts
@@ -5,10 +5,8 @@
 
 import { createSelector } from "reselect"
 import { ReduxRootState } from "App/__deprecated__/renderer/store"
-import {
-  ContactCategory,
-  ContactsState,
-} from "App/contacts/reducers/contacts.interface"
+import { ContactsState } from "App/contacts/reducers/contacts.interface"
+import { ContactCategory } from "App/contacts/dto"
 import { contactsStateSelector } from "App/contacts/selectors/contacts-state.selector"
 import { getSortedContactList } from "App/contacts/helpers/contacts.helpers"
